takeSnapshot extracted to own file
